Moved AlembicUtil class to its own file
